+++ dmr/metadata.py
@@ -391,6 +391,8 @@ class EndpointMetadata:
             to be used after auth checks.
         throttling_allow_unsafe_cache: Should this endpoint allow
             unsafe throttle Django cache backends?
+        exclude_validate_responses: Set of status codes that we don't
+            validate, even when ``validate_responses`` is enabled.
         no_validate_http_spec: Set of checks that user wants
             to disable for validation in this endpoint.
         allowed_http_methods: Set of extra HTTP methods
@@ -455,6 +457,7 @@ class EndpointMetadata:
     throttling_after_auth: tuple['SyncThrottle | AsyncThrottle', ...] | None
     throttling_allow_unsafe_cache: bool | None
 
+    exclude_validate_responses: frozenset[HTTPStatus]
     no_validate_http_spec: frozenset['HttpSpec']
     allowed_http_methods: frozenset[str]
     semantic_responses: bool